pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "OctoPrint-Fazenda3D"
version = "0.1.0"
description = "Plugin para integrar o OctoPrint com sistema de fazenda 3D"
authors = [
    {name = "Guilhemre Froes", email = "guil.henrique.fr@gmail.com"}
]
readme = {file = "README.md", content-type = "text/markdown"}
dynamic = [
    "license"
]

requires-python = ">=3.7, <4"

# any additional requirements (besides OctoPrint) should be listed here
dependencies = [
    "python-socketio[client]",
    "requests"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = [
    "octoprint_fazenda3d",
    "octoprint_fazenda3d.*"
]

[project.entry-points."octoprint.plugin"]
octoprint_fazenda3d = "octoprint_fazenda3d"

[project.urls]
Homepage = "https://github.com/GuilhermeHMFroes/octoprint-fazenda3d"

[project.optional-dependencies]
develop = [
    "go-task-bin"
]
